Initial Commit
I took a few shortcuts here as I got held up on a few issues. One, I
did not know how to reference the background image in the image file
outside of the folder where the css was located.  Second, I did not
know how to add line breaks and split up the paragraphs.